feat(#60): 스터디 삭제 API 구현
- 스터디를 삭제하는 API 구현

+++ app/src/main/java/com/ssafy/neegongnaegong/data/datasource/network/NetworkStudiesDataSource.kt
@@ -25,7 +25,7 @@ interface NetworkStudiesDataSource {
         request: UpdateStudiesRequest,
     ): Flow<Unit>
 
-    suspend fun deleteStudies(studyGroupId: Int): Flow<Unit>
+    suspend fun deleteStudies(studyGroupId: Long): Flow<Unit>
 
     suspend fun applyStudies(studyGroupId: Long): Flow<Unit>
